+++ Controls-demo/Filter_new/TumblerContainer/TumblerContainer.wml
@@ -1,5 +1,5 @@
 <div class="controlsDemo__wrapper">
         <Controls.filter:TumblerContainer source="{{_source}}">
-            <Controls.filter:Tumbler/>
+            <Controls.toggle:Tumbler/>
         </Controls.filter:TumblerContainer>
 </div>
